refactor: remove deprecated PITR test cases and utility functions; streamline restore logic
