+++ ui/apps/everest/src/modals/restore-db-modal/restore-pitr.utils.test.ts
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { PitrStorageStatus } from './restore-pitr.types';
-import { getPitrWindow, toRestoreDateISO } from './restore-pitr.utils';
+import { toRestoreDateISO } from './restore-pitr.utils';
 
 describe('toRestoreDateISO', () => {
   it('emits an offset-carrying UTC timestamp without milliseconds', () => {
@@ -27,48 +26,5 @@ describe('toRestoreDateISO', () => {
     expect(toRestoreDateISO(new Date('2026-07-29T12:15:00+02:00'))).toBe(
       '2026-07-29T10:15:00Z'
     );
-  });
-});
-
-describe('getPitrWindow', () => {
-  it('returns a usable window when Available with both bounds', () => {
-    const status: PitrStorageStatus = {
-      state: 'Available',
-      earliestRestorableTime: '2026-07-29T00:00:00Z',
-      latestRestorableTime: '2026-07-29T12:00:00Z',
-      message: 'ok',
-    };
-
-    expect(getPitrWindow(status)).toEqual({
-      available: true,
-      earliest: new Date('2026-07-29T00:00:00Z'),
-      latest: new Date('2026-07-29T12:00:00Z'),
-      message: 'ok',
-    });
-  });
-
-  it('is unavailable (with the message) when the state is Unavailable', () => {
-    const status: PitrStorageStatus = {
-      state: 'Unavailable',
-      message: 'no successful backup yet',
-    };
-
-    expect(getPitrWindow(status)).toEqual({
-      available: false,
-      message: 'no successful backup yet',
-    });
-  });
-
-  it('is unavailable when Available but a bound is missing', () => {
-    const status: PitrStorageStatus = {
-      state: 'Available',
-      latestRestorableTime: '2026-07-29T12:00:00Z',
-    };
-
-    expect(getPitrWindow(status)).toEqual({ available: false });
-  });
-
-  it('is unavailable when no status is reported', () => {
-    expect(getPitrWindow(undefined)).toEqual({ available: false });
   });
 });
